🐛 Fix PostgreSQL bigint type error in receipt matching
Cast tolerance calculation to int to prevent decimal values in BETWEEN clause. The 10% tolerance calculation was producing floats (e.g., 265089.6), which PostgreSQL's bigint column type cannot accept. Explicit int casting ensures integer bounds.

+++ app/Jobs/Data/Receipt/FindReceiptForTransactionJob.php
@@ -46,7 +46,7 @@ class FindReceiptForTransactionJob implements ShouldQueue
                 ->whereBetween('time', [$startTime, $endTime])
                 ->where(function ($query) {
                     // Amount within ±10% of transaction
-                    $tolerance = $this->transactionEvent->value * 0.1;
+                    $tolerance = (int) ($this->transactionEvent->value * 0.1);
                     $query->whereBetween('value', [
                         max(0, $this->transactionEvent->value - $tolerance),
                         $this->transactionEvent->value + $tolerance,
